requirements.txt
fastapi
uvicorn
faiss-cpu
sentence-transformers
transformers
pydantic
